Data cleaning and exploratory analysis of Ebay car sales data
The aim of this project is to clean the Ebay car sales data and analyze the included used car listings using the programming language Python. We will be using pandas and numpy module for our project.

We will be working on a dataset of used cars from eBay Kleinanzeigen, a classifieds section of the German eBay website. Dataset file: 'autos.csv'